Move API and types to dashboard
