+++ packages/frinx-dashboard/src/helpers/types/uniflow-type-guards.ts
@@ -1,15 +1,10 @@
-import { Workflow, TaskDefinition, Queue, WorkflowPayload, EventListener } from '../helpers/types';
+import { Workflow, TaskDefinition, Queue, WorkflowPayload, EventListener } from './uniflow-types';
 
 // TODO more strict in the future
 
 export function isWorkflow(workflow: unknown): workflow is Workflow {
   if (workflow !== null && typeof workflow === 'object') {
-    return (
-      'name' in workflow! &&
-      typeof workflow['name'] === 'string' &&
-      'version' in workflow! &&
-      typeof workflow['version'] === 'number'
-    );
+    return 'name' in workflow! && typeof workflow['name'] === 'string';
   }
 
   return false;
@@ -17,12 +12,7 @@ export function isWorkflow(workflow: unknown): workflow is Workflow {
 
 export function isTaskDefinition(taskDefinition: unknown): taskDefinition is TaskDefinition {
   if (taskDefinition !== null && typeof taskDefinition === 'object') {
-    return (
-      'name' in taskDefinition! &&
-      typeof taskDefinition['name'] === 'string' &&
-      'timeoutPolicy' in taskDefinition! &&
-      typeof taskDefinition['timeoutPolicy'] === 'number'
-    );
+    return 'name' in taskDefinition! && typeof taskDefinition['name'] === 'string';
   }
 
   return false;
@@ -34,12 +24,7 @@ export function isArrayTypeOf<T>(array: unknown, testFunc: (value: unknown) => v
 
 export function isEventListener(eventListener: unknown): eventListener is EventListener {
   if (eventListener !== null && typeof eventListener === 'object') {
-    return (
-      'name' in eventListener! &&
-      typeof eventListener['name'] === 'string' &&
-      'event' in eventListener! &&
-      typeof eventListener['event'] === 'string'
-    );
+    return 'name' in eventListener! && typeof eventListener['name'] === 'string';
   }
 
   return false;
@@ -47,12 +32,7 @@ export function isEventListener(eventListener: unknown): eventListener is EventL
 
 export function isQueue(queue: unknown): queue is Queue {
   if (queue !== null && typeof queue === 'object') {
-    return (
-      'queueName' in queue! &&
-      typeof queue['queueName'] === 'string' &&
-      'qsize' in queue! &&
-      typeof queue['qsize'] === 'number'
-    );
+    return 'queueName' in queue! && typeof queue['queueName'] === 'string';
   }
 
   return false;
